+++ CHANGES.rst
@@ -16,7 +16,17 @@
   .. note:: If you receive "Socket receive buffer full" errors, you
             are likely experiencing `this issue <https://github.com/esnme/ultramysql/issues/34>`_ in ultramysql and
             will need a patched version, such as the one provided in
-            `this pull request <https://github.com/esnme/ultramysql/pull/61>`_.
+            `this pull request
+            <https://github.com/esnme/ultramysql/pull/61>`_.
+- Compression of local persistent cache files has been disabled by
+  default (but there is still an option to turn it back on).
+  Operational experience showed that it didn't actually save that much
+  disk space, while substantially slowing down the reading and writing
+  process (2-4x).
+- Writing persistent cache files to disk is substantially (3x) faster for
+  any Python version less than 3.4 (including 2.7). However, it now
+  requires an amount of memory equal to the local cache size.
+
 
 2.0.0b4 (2016-07-17)
 ====================